Fixed pointer-to-int cast for 64-bit
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@17358 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: fcad256c3c268c2c5ca5d6b9097d1f1ac3917e86

+++ vision/stereo_utils/src/lowlevel.cpp
@@ -259,7 +259,7 @@ PyObject *buffer(PyObject *self, PyObject *args)
   if (!PyArg_ParseTuple(args, "i", &size))
     return NULL;
   char *original = new char[size + 16];
-  char *ptr = (char*)(((int)original + 15) & ~15);
+  char *ptr = (char*)(((int)(long)original + 15) & ~15);
   return Py_BuildValue("NL", PyBuffer_FromReadWriteMemory(ptr, size), (long long)original);
 }
